+++ backend/canisters/community/impl/src/updates/c2c_unfreeze_community.rs
@@ -1,5 +1,5 @@
 use crate::activity_notifications::handle_activity_notification;
-use crate::guards::caller_is_group_index_or_local_group_index;
+use crate::guards::caller_is_group_index_or_local_user_index;
 use crate::model::events::CommunityEventInternal;
 use crate::{RuntimeState, mutate_state, run_regular_jobs};
 use canister_api_macros::update;
@@ -7,7 +7,7 @@ use canister_tracing_macros::trace;
 use community_canister::c2c_unfreeze_community::{Response::*, *};
 use types::{EventWrapper, GroupUnfrozen, Timestamped};
 
-#[update(guard = "caller_is_group_index_or_local_group_index", msgpack = true)]
+#[update(guard = "caller_is_group_index_or_local_user_index", msgpack = true)]
 #[trace]
 async fn c2c_unfreeze_community(args: Args) -> Response {
     run_regular_jobs();