+++ backend/canisters/group/impl/src/queries/c2c_name_and_members.rs
@@ -1,10 +1,10 @@
 use crate::RuntimeState;
-use crate::guards::caller_is_group_index_or_local_group_index;
+use crate::guards::caller_is_group_index_or_local_user_index;
 use crate::read_state;
 use canister_api_macros::query;
 use group_canister::c2c_name_and_members::{Response::*, *};
 
-#[query(guard = "caller_is_group_index_or_local_group_index", msgpack = true)]
+#[query(guard = "caller_is_group_index_or_local_user_index", msgpack = true)]
 fn c2c_name_and_members(_args: Args) -> Response {
     read_state(c2c_name_and_members_impl)
 }