+++ backend/canisters/proposals_bot/impl/src/lifecycle/post_upgrade.rs
@@ -1,12 +1,14 @@
 use crate::Data;
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
+use crate::model::nervous_systems::UserSubmittedProposalResult;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
 use proposals_bot_canister::post_upgrade::Args;
 use stable_memory::get_reader;
 use tracing::info;
+use types::CanisterId;
 use utils::cycles::init_cycles_dispenser_client;
 
 #[post_upgrade]
@@ -15,8 +17,17 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
+
+    data.nervous_systems.push_decided_user_submitted_proposal(
+        CanisterId::from_text("icx6s-lyaaa-aaaaq-aaeqa-cai").unwrap(),
+        UserSubmittedProposalResult {
+            proposal_id: 9,
+            user_id: CanisterId::from_text("ss6ab-3qaaa-aaaac-aq4wa-cai").unwrap().into(),
+            adopted: true,
+        },
+    );
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);