+++ backend/canisters/online_users/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.1722](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1722-online_users)] - 2025-05-01
+
 ### Changed
 
 - Increase timeout of bounded-wait calls to 5 minutes ([#7730](https://github.com/open-chat-labs/open-chat/pull/7730))